+++ NEWS.md
@@ -1,6 +1,7 @@
 # aftables 1.0.2.9000 (development version)
 
  Major update of package back-end to use openxlsx2 functions to build and format workbook (#70).
+ Bug fix: 
 
 # aftables 1.0.2